requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
langchain-pinecone==0.2.11
pinecone>=3.2.2
python-docx==0.8.11
langchain-text-splitters==0.3.8
langchain-community==0.3.27
langchain-core>=0.3.34
langchain-ollama==0.3.6
nltk==3.8.1
requests==2.31.0
pydantic==2.11.7
python-dotenv==1.0.0
aiofiles==24.1.0
pydantic-settings>=2.4.0
pydantic_core==2.33.2
pyreadline3==3.5.4
slowapi==0.1.9
streamlit>=1.47.1